Fixed bug in empirical MCF and numdifftools method
Somewhere along the line, empirical_mean_cumulative_function had an
important line commented out and was producing the wrong estimates
in some cases (one of the examples in the reliabilty_and_mle_examples
was incorrect). It has now been fixed.

I also added the flexibility to pass keyword arguments to the
numdifftools Hessian and Gradient methods. This is useful for
setting method='complex' in some cases to get a better estimate.
However, this method selection sometimes results in NaNs, so
the default is to not use it.